refac: move PopulationGraph component files to PopulationGraph folder
